api_client_extensionを別ファイルに切り出し
