+++ task_04/lib/utility/api_client.dart
@@ -1,24 +1,4 @@
-import 'package:dio/dio.dart';
-import 'package:task_04/utility/environment_variables.dart';
-
-class APIClientExtension {
-  static BaseOptions options = BaseOptions(
-      baseUrl: EnvironmentVariables.connpassBaseURL,
-      validateStatus: (status) {
-        return status! < 300;
-      });
-
-  Dio dio = Dio(options);
-
-  APIClientExtension() {
-    dio.interceptors.addAll([
-      LogInterceptor(
-        requestBody: true,
-        responseBody: true,
-      ),
-    ]);
-  }
-}
+import 'package:task_04/utility/api_client_extension.dart';
 
 class APIClient extends APIClientExtension {
   // イベント一覧を取得を取得
